Crate rome_js_analyze

Modules

globals

This module tracks all globals variables

utils

Macros

use_block_statements_diagnostic

use_block_statements_replace_body

Enums

RuleError

Series of errors encountered when running rules on a file

Functions

analyze

Run the analyzer on the provided root: this process will use the given filter to selectively restrict analysis to specific rules / a specific source range, then call emit_signal when an analysis rule emits a diagnostic or action

analyze_with_inspect_matcher

Run the analyzer on the provided root: this process will use the given filter to selectively restrict analysis to specific rules / a specific source range, then call emit_signal when an analysis rule emits a diagnostic or action. Additionally, this function takes a inspect_matcher function that can be used to inspect the “query matches” emitted by the analyzer before they are processed by the lint rules registry

metadata

Return the static MetadataRegistry for the JS analyzer rules

visit_registry