Function rome_js_analyze::analyze_with_inspect_matcher

pub fn analyze_with_inspect_matcher<'a, V, F, B>(
    file_id: FileId,
    root: &LanguageRoot<JsLanguage>,
    filter: AnalysisFilter<'_>,
    inspect_matcher: V,
    options: &'a AnalyzerOptions,
    emit_signal: F
) -> Option<B> where
    V: FnMut(&MatchQueryParams<'_, '_, JsLanguage>) + 'a,
    F: FnMut(&dyn AnalyzerSignal<JsLanguage>) -> ControlFlow<B> + 'a,
    B: 'a, 

Run the analyzer on the provided root: this process will use the given filter to selectively restrict analysis to specific rules / a specific source range, then call emit_signal when an analysis rule emits a diagnostic or action. Additionally, this function takes a inspect_matcher function that can be used to inspect the “query matches” emitted by the analyzer before they are processed by the lint rules registry