Trait rome_js_analyze::utils::rename::RenameSymbolExtensions

pub trait RenameSymbolExtensions {
    fn rename_node_declaration(
        &mut self,
        model: &SemanticModel,
        node: impl RenamableNode,
        new_name: &str
    ) -> bool;

    fn rename_node_declaration_with_retry<S, I>(
        &mut self,
        model: &SemanticModel,
        node: impl RenamableNode + Clone,
        candidates: I
    ) -> bool
    where
        S: AsRef<str>,
        I: Iterator<Item = S>,
    { ... }
    fn rename_any_renamable_node(
        &mut self,
        model: &SemanticModel,
        node: JsAnyRenamableDeclaration,
        new_name: &str
    ) -> bool { ... }
}

Required Methods

fn rename_node_declaration(
&mut self,
model: &SemanticModel,
node: impl RenamableNode,
new_name: &str
) -> bool

Rename the binding and all its references to “new_name”.

Provided Methods

fn rename_node_declaration_with_retry<S, I>(
&mut self,
model: &SemanticModel,
node: impl RenamableNode + Clone,
candidates: I
) -> bool where
S: AsRef<str>,
I: Iterator<Item = S>,

Rename a symbol using the new name from the candidates iterator until the first success.

A usual use case is to append a suffix to a variable name.

let new_name = "new_name";
let candidates = (2..).map(|i| format!("{}{}", new_name, i).into());
let candidates = once(Cow::from(new_name)).chain(candidates);
batch.try_rename_node_declaration_until_success(model, node, candidates);

fn rename_any_renamable_node(
&mut self,
model: &SemanticModel,
node: JsAnyRenamableDeclaration,
new_name: &str
) -> bool

Rename the binding and all its references to “new_name”.