Trait rome_js_analyze::utils::ToCamelCase

pub trait ToCamelCase {
    fn to_camel_case(&self) -> Cow<'_, str>;
}

Required Methods

fn to_camel_case(&self) -> Cow<'_, str>

Return the camel case form of the input parameter. If it is already in camel case, nothing is done.

This method do not address abbreviations and acronyms.